requirements.txt
streamlit
rembg
torch
opencv-python
Pillow
numpy
scikit-learn
cohere
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
